+++ src/main/java/com/marginallyclever/ro3/apps/viewport/viewporttool/SelectionTool.java
@@ -1,4 +1,4 @@
-package com.marginallyclever.ro3.apps.viewport.viewporttools;
+package com.marginallyclever.ro3.apps.viewport.viewporttool;
 
 import com.jogamp.opengl.GL3;
 import com.marginallyclever.convenience.Ray;
